add valdators to user

+++ src/features/user/user.controller.ts
@@ -3,6 +3,9 @@ import * as userService from './user.service'
 import IUser from "./user.interface";
 import { Types } from "mongoose";
 import { IUpdateData } from "../helpers/update.interface";
+import { body, query, validationResult } from "express-validator";
+import { isUserName, isEmail, isFirstName, isLastName, isPassword, isUserIdParams, handelValidation } from "./user.validators";
+
 async function getUsersController(req: Request, res: Response): Promise<void> {
     try{
         const users = await userService.getUsersService()
@@ -15,6 +18,11 @@ async function getUsersController(req: Request, res: Response): Promise<void> {
 
 async function getUserController(req: Request, res: Response): Promise<void>{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const id = req.params.user_id
         const user = await userService.getUserService(new Types.ObjectId(id))
         res.json(user)
@@ -24,8 +32,13 @@ async function getUserController(req: Request, res: Response): Promise<void>{
 }
 
 async function createUserController(req: Request, res: Response): Promise<void>{
-    try{
+    try{  
         const reqBody = req.body;
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const userData: IUser = {
             username: reqBody.username,
             email: reqBody.email,
@@ -42,6 +55,11 @@ async function createUserController(req: Request, res: Response): Promise<void>{
 
 async function updateUserController(req: Request, res: Response): Promise<void>{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const reqBody = req.body;
         const {user_id} = req.params;
         const data: IUpdateData = {
@@ -51,7 +69,7 @@ async function updateUserController(req: Request, res: Response): Promise<void>{
             last_name: reqBody.last_name,
         }
         const user = await userService.updateUserService(new Types.ObjectId(user_id), data);
-        res.sendStatus(200);
+        res.sendStatus(204);
     }catch(err: any){
         res.status(400).json(err.message);
     }
@@ -59,6 +77,11 @@ async function updateUserController(req: Request, res: Response): Promise<void>{
 
 async function deleteUserController(req: Request, res: Response): Promise<void>{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const isDeleted = await userService.deleteUserService(new Types.ObjectId(req.params.user_id));
         if (isDeleted){
             res.sendStatus(204);
@@ -69,11 +92,46 @@ async function deleteUserController(req: Request, res: Response): Promise<void>{
         res.status(400).json(err.message);
     }
 }
-
+function validate(method: string) {
+    switch (method) {
+        case 'createUser': {
+            return [
+                isUserName(),
+                isEmail(),
+                isFirstName(),
+                isLastName(),
+                isPassword()
+            ];
+        }
+        case 'updateUser': {
+            return [
+                isUserName(),
+                isEmail(),
+                isFirstName(),
+                isLastName(),
+                isUserIdParams()
+            ];
+        }
+        case 'getUser': {
+            return [
+                isUserIdParams()
+            ];
+        
+        }
+        case 'deleteUser': {
+            return [
+                isUserIdParams()
+            ];
+        }
+    }
+    return [body()];
+}
 export {
     getUserController,
     getUsersController,
     updateUserController,
     deleteUserController,
-    createUserController
+    createUserController,
+    validate
 }
+
